Use autotools_clang_conda

+++ recipes/zimg/build.sh
@@ -2,7 +2,9 @@ set -ex
 
 ./autogen.sh
 ./configure --prefix="$PREFIX" --enable-shared
+[[ "$target_platform" == "win-64" ]] && patch_libtool
 make -j${CPU_COUNT}
 make install
 
-rm "$PREFIX"/lib/libzimg.a
+# only when not on windows
+[[ "$target_platform" != "win-64" ]] && rm "$PREFIX"/lib/libzimg.a
